refactor: Extract supervisor actions into separate modules
- Split supervisor actions into dedicated files
- Create fault handler module for error handling
- Move process states to constants module
- Extract stop/kill logic into standalone functions
- Reorganize supervisor executor module
